requirements.txt
flask~=2.3.2
dlib==19.24.0
face-recognition==1.3.0
numpy==1.24.2
Pillow==9.4.0
google-cloud-storage
google-cloud-firestore
flask-login
tabulate
